requirements.txt
Django>=3.2
pandas
numpy
matplotlib
statsmodels>=0.13.0
joblib
djangorestframework
